requirements.txt
beautifulsoup4~=4.12
discord.py~=2.4
feedparser~=6.0
html2markdown~=0.1
html5lib~=1.1
humanize~=4.10
jinja2~=3.1
requests~=2.32
prettytable~=3.11
hashids~=1.3
pillow~=10.4
Wand~=0.6.13
hjson~=3.1
aiohttp~=3.10
python-dotenv~=1.0